+++ packages/emoji-picker/src/Group.tsx
@@ -1,4 +1,5 @@
 import React, { useCallback, useContext } from 'react';
+import { GroupKey as BaseGroupKey } from 'emojibase';
 import camelCase from 'lodash/camelCase';
 import { GROUP_KEY_COMMONLY_USED } from './constants';
 import Context from './Context';
@@ -13,7 +14,7 @@ export interface GroupProps {
 }
 
 export default function Group({ active, children, commonMode, group, onSelect }: GroupProps) {
-  const { classNames, messages } = useContext(Context);
+  const { classNames, emojiData, messages } = useContext(Context);
   const key = camelCase(group === GROUP_KEY_COMMONLY_USED ? commonMode : group);
   const className = [classNames.group];
 
@@ -32,7 +33,7 @@ export default function Group({ active, children, commonMode, group, onSelect }:
   return (
     <button
       className={className.join(' ')}
-      title={messages[key]}
+      title={emojiData?.GROUPS_BY_KEY?.[group as BaseGroupKey] || messages[key]}
       type="button"
       onClick={handleClick}
     >