+++ packages/emoji/src/EmojiDataManager.ts
@@ -4,8 +4,11 @@ import {
   fromCodepointToUnicode,
   fromHexcodeToCodepoint,
   generateEmoticonPermutations,
+  GroupKey,
   Hexcode,
   Locale,
+  MetadataDataset,
+  SubgroupKey,
   TEXT,
 } from 'emojibase';
 import { CanonicalEmoji } from './types';
@@ -19,13 +22,17 @@ export function resetInstances() {
 }
 
 export default class EmojiDataManager {
-  EMOJIS: { [hexcode: string]: CanonicalEmoji } = {};
+  EMOJIS: Record<string, CanonicalEmoji> = {};
 
-  EMOTICON_TO_HEXCODE: { [emoticon: string]: Hexcode } = {};
+  EMOTICON_TO_HEXCODE: Record<string, Hexcode> = {};
 
-  SHORTCODE_TO_HEXCODE: { [shortcode: string]: Hexcode } = {};
+  SHORTCODE_TO_HEXCODE: Record<string, Hexcode> = {};
 
-  UNICODE_TO_HEXCODE: { [unicode: string]: Hexcode } = {};
+  UNICODE_TO_HEXCODE: Record<string, Hexcode> = {};
+
+  GROUPS_BY_KEY: Partial<Record<GroupKey, string>> = {};
+
+  SUBGROUPS_BY_KEY: Partial<Record<SubgroupKey, string>> = {};
 
   data: CanonicalEmoji[] = [];
 
@@ -134,11 +141,21 @@ export default class EmojiDataManager {
       // Flatten and package skins as well
       if (packagedEmoji.skins) {
         packagedEmoji.skins.forEach((skin) => {
-          this.flatData.push(this.packageEmoji(skin));
+          this.flatData.push(skin);
         });
       }
     });
 
     return this.data;
+  }
+
+  parseMessageData(data: MetadataDataset) {
+    data.groups.forEach((group) => {
+      this.GROUPS_BY_KEY[group.key as GroupKey] = group.message;
+    });
+
+    data.subgroups.forEach((group) => {
+      this.SUBGROUPS_BY_KEY[group.key as SubgroupKey] = group.message;
+    });
   }
 }