+++ packages/emoji-picker/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - Removed `emojibase` as a peer requirement.
 
+#### 🚀 Updates
+
+- Updated to support Emoji v13, which includes 117 new emojis.
+- Updated to support different shortcodes using pre-defined presets.
+- Added `shortcodes` prop to `Picker`.
+
 #### 📦 Dependencies
 
 - Updated `emojibase` to v5.